Perf tests: Clone local checkout of repo for pull requests
When running the performance test suite on individual commits in a
PR we're fetching from the upstream remote repository. The remote
fetching introduces latency and variability in the test runs that's
dependendant on the networking behavior during the test setup time.

In this patch we're adding a local remote if one appears to exist
during the test run and fetching from it instead of fetching from
the network. This should streamline the test setup and avoid needing
to download data that's already available on the test runner container.