Performance Test: Reuse tests-branch build if possible
When running the performance test suite we currently build three copies
of the plugin: one for each branch under test and also one for the branch
which provides the actual test suite.

In this patch we're checking first if one of the branches under test is
the same as that tests branch, and if so, re-using the built files from
that to avoid the additional approximately five minutes of building.